add .env.example and move models.py to models directory
