+++ .env.example
@@ -1,0 +1,11 @@
+API_V1_STR=/api/v1
+SECRET_KEY=YOUR_SECRET_KEY
+ACCESS_TOKEN_EXPIRE_MINUTES=11520
+ENVIRONMENT=local
+BACKEND_CORS_ORIGINS="*"
+PROJECT_NAME=FastAPI
+POSTGRES_SERVER=localhost
+POSTGRES_PORT=5432
+POSTGRES_USER=root
+POSTGRES_PASSWORD=
+POSTGRES_DB=
