+++ app/models/token.py
@@ -1,0 +1,11 @@
+from sqlmodel import SQLModel
+
+
+class Token(SQLModel):
+    access_token: str
+    token_type: str = "bearer"
+
+
+# Contents of JWT token
+class TokenPayload(SQLModel):
+    sub: str | None = None
