+++ app/tests/conftest.py
@@ -5,7 +5,7 @@ from fastapi.testclient import TestClient
 from sqlmodel import Session, delete
 
 from app.core.config import settings
-from app.core.db import engine, init_db
+from app.core.db import engine
 from app.main import app
 from app.models import Item, User
 from app.tests.utils.user import authentication_token_from_email
@@ -15,7 +15,6 @@ from app.tests.utils.utils import get_superuser_token_headers
 @pytest.fixture(scope="session", autouse=True)
 def db() -> Generator[Session, None, None]:
     with Session(engine) as session:
-        init_db(session)
         yield session
         statement = delete(Item)
         session.execute(statement)
